+++ urlhub.js
@@ -109,15 +109,8 @@ var prototype = {
     path =  parsed.pathname;
 
     while( i < candidates.length && !found ){
-
       c = candidates[i];
-      found = c.regex.exec( path );
-      if( found ){
-        found = {
-          id: c.id, cb: c.cb, params: found.slice(1)
-        };
-      }
-      else if( c.childRegex ){
+      if( c.childRegex ){
         //console.log( 'failed', c.regex, path );
         found = c.childRegex.exec( path );
         if( found ){
@@ -130,6 +123,13 @@ var prototype = {
             found = false;
           }
         }
+      }
+      
+      found = c.regex.exec( path );
+      if( found ){
+        found = {
+          id: c.id, cb: c.cb, params: found.slice(1)
+        };
       }
 
       if( !found ){
@@ -163,18 +163,17 @@ var prototype = {
   // Routing methods
   start: function(){
     var me = this;
-    this.strategy.onChange( function( location ){
-      var match = me.runOnBeforeChange( me.match( location ) );
-      if( !match ) return;
+    this.strategy.onChange( function(){
+      var change = me.checkChange();
+      if( !change.next ) return;
 
-      var nextLocation = match.pathname + match.search + match.hash;
-      if( location !== nextLocation ){
-        me.strategy.replace( nextLocation );
+      if( change.current !== change.next ){
+        me.strategy.replace( change.next );
       }
       else {
-        me.location = match;
+        me.location = change.nextLocation;
         me.cbs.forEach( function( cb ){
-          cb( match );
+          cb( change.nextLocation );
         });
       }
     });
@@ -184,6 +183,18 @@ var prototype = {
   },
   stop: function(){
     this.strategy.onChange( function(){} );
+  },
+  refresh: function(){
+    var change = this.checkChange();
+    change.next && change.current !== change.next && this.strategy.replace( change.next );
+  },
+  checkChange: function(){
+    var current = this.strategy.getLocation(),
+      nextLocation = this.runOnBeforeChange( this.match(current) ),
+      next = nextLocation && (nextLocation.pathname + nextLocation.search + nextLocation.hash)
+    ;
+
+    return {current:current, next:next, nextLocation:nextLocation};
   },
   runOnBeforeChange: function( match ){
     var me = this;